migrate subscribed tests
